refactor(logging): extraire RequestIdMiddleware et filtres dans middleware.py
